chore: refactored into one file to share typing variables and regions
