+++ pipe_operator/python_flow/__init__.py
@@ -1,7 +1,4 @@
-from .base import PipeEnd, PipeStart
-from .pipes.asynchronous import AsyncPipe
-from .pipes.basics import Pipe, Tap
-from .pipes.threads import ThreadPipe, ThreadWait
+from .pipes import AsyncPipe, Pipe, PipeEnd, PipeStart, Tap, ThreadPipe, ThreadWait
 
 __all__ = [
     "AsyncPipe",
